refactor(etl): organise csv utils to a seperate module and clean up processCatalogETL
